Close the CellTag codec gap, gate the axiom claim, and prove the budget bound
Three defects found by reading the code against what it claims, plus
the Solidity warning set.

**`CellTag`'s decoder covered 7 of 15 constructors.**
`Encoding/KernelStep.lean`'s `CellTag.encode` emits tags 0..14 across
every constructor; `CellTag.decode` handled 0..6 and answered
`invalidConstructorIndex` to the rest.  The eight it refused are not
exotic — every action writes an `.epochBudget` cell (13) and every
frontier leads with `.budgetPolicy` (14), so the gap covered the
majority of real bundles.

Nothing caught it because the module's only theorems were
`*_encode_deterministic`, i.e. `t₁ = t₂ → encode t₁ = encode t₂`,
which holds of every function and therefore cannot witness a decoder
at all; and because no test exercised `CellTag.decode`.  The eight
arms are added, and `cellTag_roundtrip` now states the property that
was missing, carrying `CellTag.fieldsBounded` — a real conjunct only
for the two bridge-set cells, whose ids are bare `Nat` rather than
`UInt64`.  New suite `encoding-kernelstep` sweeps every constructor
off an arity-pinned list, with an unknown-tag rejection and a
pairwise-distinctness control so it cannot pass vacuously.

**"No custom axioms (ABSOLUTE)" had no mechanical gate.**  Six audit
binaries ship beside that claim and none checked it; `count_sorries`
says so itself ("A full check would invoke Lean's elaborator and
inspect `sorryAx` axiom usage").  `Test/AxiomFootprint.lean` adds
`#assert_canonical_axioms`, a command — so a violation is a build
error, and CI's existing `lake build` gates it with no new step — run
over the headline-theorem table.  Verified to fail on both a custom
`axiom` and a transitive `sorryAx` written `first | sorry`, a spelling
`count_sorries`' four patterns do not match.

**The epoch-budget growth bound was asserted, unproved, and wrong.**
`BoundsReachable.lean` said the 8-byte head is safe because "a budget
[advances] by at most MAX_TOPUP_BUDGET_PER_ACTION", and that the
argument "is stated here where it can be checked rather than left in a
comment".  Only the nonce half had theorems.  The budget claim is
false twice over: `normalise` floors a stale cell at the policy's
`freeTier` before the credit lands, so one step can lift a balance
from 0 to `freeTier + amount` however small the grant, and `freeTier`
is an unbounded `Nat` that `mkBounded` does not clamp; and
`depositWithFee`'s `budgetGrant` never meets the cap, since
`depositWithFee_signerCheck` constrains the signer and not the amount.

The true per-step bound is `max stored freeTier + grant`, now proved:
`storedBalance_topUp_le` / `_consume_le` over `storedBalance`, the raw
number `budgetCellValue` encodes — `currentBudget` folds the free-tier
floor in and so cannot be transported to the cell.  The docstring
states this, including the consequence: at the ceiling Lean truncates
while Solidity reverts (`CBEValueTooWide`), so the stacks disagree
rather than merely losing precision.

**Solidity warnings.**  `forge build` emitted one solc warning in
project source and eleven forge-lint warnings; project-source warnings
are now zero (only vendored forge-std / OpenZeppelin remain).
`StepWrites._appendUniform`'s `at` parameter is renamed (`at` becomes
a reserved keyword).  `MAX_ACTION_KIND` is typed `uint8` — the width
both consumers already take — which removes two truncating casts
rather than annotating them.  The `1 << d` shifts become
`uint256(1) << d` / `uint8(1) << …`.  The three casts that are
genuinely identities (`& 0xFF`, and `feeBps` bounded by the assertion
above it) carry a justification naming why.

lake test 3241 passed; cargo test 2378 passed; forge test 929 passed.

+++ solidity/src/contracts/KnomosisStepVMRoot.sol
@@ -119,7 +119,15 @@ contract KnomosisStepVMRoot {
     uint256 public constant MAX_CELL_OPENINGS = 32;
 
     /// @notice The highest frozen `Action` dispatcher index.
-    uint256 internal constant MAX_ACTION_KIND = 24;
+    ///
+    /// @dev    Typed `uint8` because that is the width the action kind
+    ///         has on the wire (`actionKindByte`) and the width both
+    ///         consumers — `StepWrites.isAdjudicable` and
+    ///         `StepWrites.deriveWriteSet` — accept.  Declaring it
+    ///         `uint256` forced a truncating `uint8(k)` cast at every
+    ///         call site; typing it here removes the cast rather than
+    ///         annotating it as safe.
+    uint8 internal constant MAX_ACTION_KIND = 24;
 
     /// @notice Field-buffer length `widestFrontier` probes with —
     ///         comfortably past the longest layout `actionFieldsForL1`
@@ -197,10 +205,10 @@ contract KnomosisStepVMRoot {
         returns (uint256 widest)
     {
         require(probeFields.length >= PROBE_FIELD_BYTES, "ProbeFieldsTooShort");
-        for (uint256 k = 0; k <= MAX_ACTION_KIND; k++) {
-            if (!StepWrites.isAdjudicable(uint8(k))) continue;
+        for (uint8 k = 0; k <= MAX_ACTION_KIND; k++) {
+            if (!StepWrites.isAdjudicable(k)) continue;
             uint256 n =
-                StepWrites.deriveWriteSet(uint8(k), probeFields, 0, 0).length + 1;
+                StepWrites.deriveWriteSet(k, probeFields, 0, 0).length + 1;
             if (n > widest) widest = n;
         }
     }
